examples/openai_realtime_speech_to_speech/pyproject.toml
[project]
name = "openapi-realtime-speech-to-speech-example"
version = "0.1.0"
description = "Example project showing how to use OpenAI's realitme speech-to-speech model with the Stream Python SDK."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "python-dotenv>=1.0.0",
    "aiortc>=1.10.1",
    "numpy>=2.0.0",
    "openai>=1.82.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]
getstream = { workspace = true }
